More restructuring not finished
Still cannot get the dependent projects to find the location of the
libraries as no good way to speficy where the library will be after
building due to shadow builds. I need to be able to specify to either
install it to a location, or generate the name that it will be found
under.

Signed-off-by: David Post <david@ohyonghao.com>