Got the kernel running with output
Not sure why this worked before, perhaps it was the bpp was given
as 4 from Qt and so things just worked there, but in the command
line version working directly with my bitmap loader there was a
size difference leading to segmentation fault.

Kernel still does not perform properly, but I have created output
files that are the expected output that I can compare with. I
will continue working to generalize the kernel to the image
parameters which it currently is not. Once that is done then it is
on to generalizing it to hardware capabilities.

Signed-off-by: David Post <david@ohyonghao.com>