Added swap_work function to Image class
This allows us to easily flip the buffer to apply another kernel.
We guarantee that after a kernel is ran that result() gives the
result, but data() may not be the original image anymore. May
rethink the need for data and result in the future and just hide
this behind data().

Signed-off-by: David Post <david@ohyonghao.com>